delete all mappers
